backend/services/base/python/requirements.txt
flask
flask-restful
flask-sqlalchemy
sqlalchemy-utils
passlib
requests
uwsgi
python-jose
pytest
psycopg2